feat: users can only update and delete their own blog posts.
new: docker config added

+++ posts/mutations.py
@@ -2,6 +2,7 @@ import graphene
 from graphql import GraphQLError
 from posts.forms import PostForm
 from posts.models import Post
+from django.contrib.auth.models import User
 
 
 class CreatePost(graphene.Mutation):
@@ -14,12 +15,17 @@ class CreatePost(graphene.Mutation):
     success = graphene.Boolean()
 
     def mutate(self, info, title, sub_title, body, category_id, **kwargs):
+        user = info.context.user
+        if not user.is_authenticated:
+            raise GraphQLError("You must be logged in to create a post")
+
         form = PostForm(
             {
                 "title": title,
                 "sub_title": sub_title,
                 "body": body,
                 "category": category_id,
+                "author": user,
             }
         )
         if form.is_valid():
@@ -40,10 +46,15 @@ class UpdatePost(graphene.Mutation):
     success = graphene.Boolean()
 
     def mutate(self, info, id, title, sub_title, body, category_id, **kwargs):
+        user = info.context.user
         try:
             existing_post = Post.objects.get(id=id)
         except Post.DoesNotExist:
             raise GraphQLError("Post not found")
+
+        # check if requesting user is the author of the post
+        if existing_post.author != user:
+            raise GraphQLError("You are not authorized to update this post")
 
         form = PostForm(
             {
@@ -51,6 +62,7 @@ class UpdatePost(graphene.Mutation):
                 "sub_title": sub_title,
                 "body": body,
                 "category": category_id,
+                "author": user,
             },
             instance=existing_post,
         )
@@ -68,10 +80,14 @@ class DeletePost(graphene.Mutation):
     success = graphene.Boolean()
 
     def mutate(self, info, id, **kwargs):
+        user = info.context.user
         try:
             existing_post = Post.objects.get(id=id)
         except Post.DoesNotExist:
             raise GraphQLError("Post not found")
+        
+        if existing_post.author != user:
+            raise GraphQLError("You are not authorized to delete this post")
 
         existing_post.delete()
         return DeletePost(success=True)
